// frontend/src/api/generated/model/capacityExceeded.ts
export interface CapacityExceeded {
  message: string;
  room_id?: number;
  room_number?: string;
  capacity: number;
  beds: number;
  /** @minimum 0 */
  free_places: number;
}