// frontend/src/api/generated/model/entryNotPermittedError.ts
export interface EntryNotPermittedError {
  message: string;
  reason_code?: EntryNotPermittedErrorReasonCode;
  override_available?: boolean;
  override_requires_reason?: boolean;
}