// frontend/src/api/generated/model/quotaError.ts
export interface QuotaError {
  message: string;
  quota_scope?: QuotaErrorQuotaScope;
  quota_limit?: number;
  approved_today?: number;
  visit_date?: string;
}